+++ weatherservicelib/src/main/java/com/markosopcic/weatherservicelib/models/Current.kt
@@ -1,18 +1,18 @@
 package com.markosopcic.weatherservicelib.models
 
 data class Current(
-    val clouds: Int,
+    val clouds: Double,
     val dewPoint: Double,
     val datetimeSeconds: Int,
     val feelsLike: Double,
-    val humidity: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pressure: Double,
     val sunrise: Int,
     val sunset: Int,
     val temp: Double,
     val uvi: Double,
-    val visibility: Int,
+    val visibility: Double,
     val weather: List<Weather>,
-    val windDeg: Int,
+    val windDeg: Double,
     val windSpeed: Double
 )
