+++ weatherservicelib/src/main/java/com/markosopcic/weatherservicelib/models/Daily.kt
@@ -1,19 +1,19 @@
 package com.markosopcic.weatherservicelib.models
 
 data class Daily(
-    val clouds: Int,
+    val clouds: Double,
     val dewPoint: Double,
     val timestampSeconds: Int,
     val feelsLike: FeelsLike,
-    val humidity: Int,
-    val precipitationPercentage: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val precipitationPercentage: Double,
+    val pressure: Double,
     val rain: Double,
     val sunrise: Int,
     val sunset: Int,
     val temp: Temperature,
     val uvi: Double,
     val weather: List<Weather>,
-    val windDeg: Int,
+    val windDeg: Double,
     val windSpeed: Double
 )
