+++ weatherservicelib/src/main/java/com/markosopcic/weatherservicelib/models/Main.kt
@@ -1,8 +1,8 @@
 package com.markosopcic.weatherservicelib.models
 
 data class Main(
-    val humidity: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pressure: Double,
     val temperature: Double,
     val temperatureMax: Double,
     val temperatureMin: Double,
