+++ weatherservicelib/src/main/java/com/markosopcic/weatherservicelib/models/Minutely.kt
@@ -2,5 +2,5 @@ package com.markosopcic.weatherservicelib.models
 
 data class Minutely(
     val timestampSeconds: Int,
-    val precipitation: Int
+    val precipitation: Double
 )
